master commit

+++ opencv_common/opencv_common/cv_util/cv_string.cpp
@@ -20,7 +20,7 @@ String wcharAt(String s, int idx){
 	int p = 0;
 	const char *buff = s.c_str();
 
-	for(int i=0; i<idx && p < s.size() ; i++){
+	for(int i=0; i<idx && p < (int) s.size() ; i++){
 		if( ( buff[p] & 0x80 ) == 0x80 ) p += 2;
 		else p++;
 	}
